here lies the "impass":
- full breakout of tests into separate package will break hasAction local test helper
- need to renovate the controller tests to use KM interface
